--- htmlAnalyse/config.js
@@ -1,17 +1,0 @@
-'use strict'
-
-var config = [
-  ['&quot;', '"'],
-  ['&amp;', '&'],
-  ['&lt;', '<'],
-  ['&gt;', '>'],
-  ['&nbsp;', ' '],
-  ['&amp;', '&'],
-  ['%3A', ':'],
-  ['%3F', '?'],
-  ['%3D', '='],
-  ['%26', '&']
-]
-module.exports = {
-  config: config
-}
