+++ zhihuapi/api.js
@@ -1,8 +1,6 @@
-var GetCollection = require('./zhihu/collection.js').GetCollection
-var GetCollectionPages = require('./zhihu/collection.js').GetCollectionPages
 module.exports = {
   zhihu: {
-    GetCollection: GetCollection,
-    GetCollectionPages: GetCollectionPages
+    GetCollection: require('./zhihu/collection.js').GetCollection,
+    GetCollectionPages: require('./zhihu/collection.js').GetCollectionPages
   }
 }
